backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
httpx==0.28.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
